+++ src/App.vue
@@ -1,20 +1,32 @@
-<script setup>
-// This starter template is using Vue 3 <script setup> SFCs
-// Check out https://vuejs.org/api/sfc-script-setup.html#script-setup
-import Form from './components/Form.vue';
-</script>
-
 <template>
-  <div>
-    <a href="https://vitejs.dev" target="_blank">
-      <img src="/vite.svg" class="logo" alt="Vite logo" />
-    </a>
-    <a href="https://vuejs.org/" target="_blank">
-      <img src="./assets/vue.svg" class="logo vue" alt="Vue logo" />
-    </a>
-  </div>
-  <HelloWorld msg="Vite + Vue" />
+  <form @submit.prevent="getData">
+    <input v-model="pid" />
+    <input type="submit" value="Get Data" />
+  </form>
+  <pre>{{ formData }}</pre>
 </template>
+
+<script setup>
+import { ref } from 'vue';
+const pid = ref('');
+const formData = ref(null);
+
+const getData = async () => {
+  try {
+    const { data } = await fetch(
+      `https://sur10.dyn.konopnickiej.com/5510/?pid=${pid.value}`
+    );
+    formData.value = data.json();
+  } catch (err) {
+    console.error(err);
+  }
+};
+</script>
+<style scoped>
+.read-the-docs {
+  color: #888;
+}
+</style>
 
 <style scoped>
 .logo {
